// packages/frontend/src/api/models/ExerciseCreationVitalSigns.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ExerciseCreationVitalSigns
 */
export interface ExerciseCreationVitalSigns {
    /**
     * 
     * @type {string}
     * @memberof ExerciseCreationVitalSigns
     */
    temperature: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseCreationVitalSigns
     */
    heartRate: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseCreationVitalSigns
     */
    respiratoryRate: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseCreationVitalSigns
     */
    bloodPressure: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseCreationVitalSigns
     */
    bloodGlucose: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseCreationVitalSigns
     */
    oxygenSaturation: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseCreationVitalSigns
     */
    capillaryRefill: string;
}

/**
 * Check if a given object implements the ExerciseCreationVitalSigns interface.
 */
export function instanceOfExerciseCreationVitalSigns(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "temperature" in value;
    isInstance = isInstance && "heartRate" in value;
    isInstance = isInstance && "respiratoryRate" in value;
    isInstance = isInstance && "bloodPressure" in value;
    isInstance = isInstance && "bloodGlucose" in value;
    isInstance = isInstance && "oxygenSaturation" in value;
    isInstance = isInstance && "capillaryRefill" in value;

    return isInstance;
}

export function ExerciseCreationVitalSignsFromJSON(json: any): ExerciseCreationVitalSigns {
    return ExerciseCreationVitalSignsFromJSONTyped(json, false);
}

export function ExerciseCreationVitalSignsFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExerciseCreationVitalSigns {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'temperature': json['temperature'],
        'heartRate': json['heartRate'],
        'respiratoryRate': json['respiratoryRate'],
        'bloodPressure': json['bloodPressure'],
        'bloodGlucose': json['bloodGlucose'],
        'oxygenSaturation': json['oxygenSaturation'],
        'capillaryRefill': json['capillaryRefill'],
    };
}

export function ExerciseCreationVitalSignsToJSON(value?: ExerciseCreationVitalSigns | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'temperature': value.temperature,
        'heartRate': value.heartRate,
        'respiratoryRate': value.respiratoryRate,
        'bloodPressure': value.bloodPressure,
        'bloodGlucose': value.bloodGlucose,
        'oxygenSaturation': value.oxygenSaturation,
        'capillaryRefill': value.capillaryRefill,
    };
}